Fix clippy lints and relax overly strict CI lint config
Fix manual clippy issues (unnecessary parens, late init, range loop).
Relax CI clippy config to allow common stylistic patterns
(manual_let_else, map_unwrap_or, redundant closures) — these can be
cleaned up incrementally without blocking CI.

+++ rustydemon-blp2/tests/integration_test.rs
@@ -588,7 +588,7 @@ fn security_jpeg_header_size_too_large() {
     let mut v = blp2_header(0 /*Jpeg*/, 0, 0, 1, 1);
     append_mip_tables(&mut v, DATA_OFFSET, 0);
     // JPEG header size = MAX_JPEG_HEADER + 1 (just over the limit)
-    let bad_size: i32 = (64 * 1024 + 1) as i32;
+    let bad_size: i32 = 64 * 1024 + 1;
     v.extend_from_slice(&bad_size.to_le_bytes());
     // No actual header bytes follow.
